separate text_editor
